Restructure project's folder organization
