+++ projeto-kokama/src/index.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import Translation from './screens/Translation/index';
+
+const Root = () => {
+    return (
+        <Translation/>
+    );
+}
+
+export default Root;
