+++ projeto-kokama/src/screens/Translation/dictionary.json
