+++ projeto-kokama/src/screens/Translation/styles.ts
